feat: Add an input to set the Renovate log level
This allows to enable debug logging [1] when manually triggering the
workflow.

The `LOG_LEVEL` variable is set for the whole workflow, because
accessing `github.events.input` there does not cause an error when it
does not exist like when doing the same in a step. The workaround is
taken from [2].

[1]: https://github.com/renovatebot/github-action?tab=readme-ov-file#debug-logging
[2]: https://github.com/orgs/community/discussions/74698#discussioncomment-9009682

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>